The unittest program has been updated.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/branches/RELENG_1_0@1984 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/rtm/tests/PublisherNew/PublisherNewTests.cpp
@@ -565,7 +565,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc, (long)rtd.data);
         }
 
         //provider 側のバッファ full ではない状態でコール
@@ -598,7 +598,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc+4, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc+4, (long)rtd.data);
         }
         publisher.deactivate();
         
@@ -711,7 +711,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc, (long)rtd.data);
         }
 
         //consumer の buffer が full 状態のため、
@@ -735,7 +735,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc+8, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc+8, (long)rtd.data);
         }
         {
         cdrMemoryStream cdr;
@@ -755,7 +755,7 @@ namespace PublisherNew
 
         RTC::TimedLong rtd;
         rtd <<= data;
-        CPPUNIT_ASSERT_EQUAL((long)18, rtd.data);
+        CPPUNIT_ASSERT_EQUAL((long)18, (long)rtd.data);
         }
 
         coil::usleep(10000);
@@ -870,7 +870,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc, (long)rtd.data);
         }
 
         //provider 側のバッファ full ではない状態でコール
@@ -921,7 +921,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc+4, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc+4, (long)rtd.data);
         }
 
         //この時点で consumer 側のバッファにデータが 2 件格納されている状態
@@ -1038,7 +1038,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc, (long)rtd.data);
         }
 
         //consumer の buffer が full 状態のため、
@@ -1073,7 +1073,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc+8, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc+8, (long)rtd.data);
         }
         {
         cdrMemoryStream cdr;
@@ -1093,7 +1093,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)18, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)18, (long)rtd.data);
         }
 
 
@@ -1207,7 +1207,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc*2+1, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc*2+1, (long)rtd.data);
         }
 
         //provider 側のバッファ full ではない状態でコール
@@ -1240,7 +1240,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc*2+9, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc*2+9, (long)rtd.data);
         }
        
         coil::usleep(100000);
@@ -1354,7 +1354,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc*2+1, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc*2+1, (long)rtd.data);
         }
 
         //consumer の buffer が full 状態のため、
@@ -1380,7 +1380,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)icc*2+17, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)icc*2+17, (long)rtd.data);
         }
         {
         cdrMemoryStream cdr;
@@ -1409,7 +1409,7 @@ namespace PublisherNew
 
             RTC::TimedLong rtd;
             rtd <<= data;
-            CPPUNIT_ASSERT_EQUAL((long)27, rtd.data);
+            CPPUNIT_ASSERT_EQUAL((long)27, (long)rtd.data);
         }
 
         coil::usleep(10000);
@@ -1512,7 +1512,7 @@ namespace PublisherNew
 
         RTC::TimedLong rtd;
         rtd <<= data;
-        CPPUNIT_ASSERT_EQUAL((long)7, rtd.data);
+        CPPUNIT_ASSERT_EQUAL((long)7, (long)rtd.data);
         }
 
         coil::usleep(1000000);  // ここは長めにしないと落ちます。
